Added more unit tests for file:files and corrected the bad existing ones. Corrected bug in testdriver spec reading: wrong usage of function good() for streams. Good is set to false only after eof has been reached. Therefore the following code is wrong:     while(lFile.good()) {       lContent = (char)lFile.get();     } the following code is the correct version:     char c = (char)lFile.get();     while(lFile.good()) {       lContent &lt;&lt; c;       c = (char)lFile.get();     }

+++ test/rbkt/testdriver_common.cpp
@@ -51,7 +51,7 @@ void slurp_file (
   assert (qfile);
 
   qfile.seekg (0, std::ios::end);
-  size_t len = qfile.tellg ();
+  unsigned int len = static_cast<unsigned int> (qfile.tellg ());
   qfile.seekg (0, std::ios::beg);
 
   char *str = new char [len];
